requirements/base.txt
django==3.*
psycopg2==2.8.6
environs==9.3.1 # https://pypi.org/project/environs/#basic-usage
django-extensions==3.1.1
Werkzeug==1.0.1 # required for django-extensions
djangorestframework==3.12.2